requirements.txt
altair==5.3.0
attrs==23.2.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cvzone==1.6.1
cycler==0.12.1
filelock==3.16.1
filterpy==1.4.5
fonttools==4.55.2
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
imageio==2.36.1
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
numerize==0.12
numpy==1.26.4
opencv-contrib-python==4.10.0.82
opencv-python==4.10.0.84
packaging==24.0
pandas==2.2.2
pillow==10.3.0
plotly==5.22.0
protobuf==4.25.3
psutil==6.1.0
py-cpuinfo==9.0.0
pyarrow==16.1.0
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
scikit-image==0.24.0
scipy==1.14.1
seaborn==0.13.2
setuptools==75.6.0
six==1.16.0
smmap==5.0.1
streamlit==1.35.0
streamlit-option-menu==0.3.13
sympy==1.13.1
tenacity==8.3.0
tifffile==2024.9.20
toml==0.10.2
toolz==0.12.1
torch==2.5.1
torchvision==0.20.1
tornado==6.4
tqdm==4.67.1
typing_extensions==4.12.1
tzdata==2024.1
ultralytics==8.3.48
ultralytics-thop==2.0.13
urllib3==2.2.1
watchdog==4.0.1
wheel==0.43.0
